+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0]
+
+### Added
+
+- Add new map property mushymato.MMAP_QuestionDialogue, and new custom asset mushymato.MMAP/QuestionDialogue
+
+## [1.2.0]
+
+### Added
+
+- Add new map property mushymato.MMAP_FruitTreeCosmeticSeason
+
 ## [1.1.1]
 
 ### Added